refactor: change name from board to dashboard v2
