+++ packages/backend/src/index.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-console */
 import dotenv from 'dotenv'
 import Task from './modules/task/Task'
-import Dashboard from './modules/board/Dashboard'
+import Dashboard from './modules/dashboard/Dashboard'
 import app from './app'
 import { connectToDatabase } from './infrastructure/mongoose'
 import User from './modules/user/User'
